# backend-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "Dokumentasi API v1.0.0",
        "contact": {
            "email": "prasdimasetyo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/candidates": {
            "get": {
                "tags": [
                    "Candidates"
                ],
                "summary": "Get all candidates",
                "operationId": "c9b0e1a5e84add97746106caab9aba4f",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": [
                                    {
                                        "id": 1,
                                        "name": "John Doe",
                                        "email": "john@example.com",
                                        "phone": "123456789",
                                        "year": 2022,
                                        "skillSets": [
                                            {
                                                "id": 1,
                                                "skill_id": 1
                                            }
                                        ],
                                        "job": {
                                            "id": 1,
                                            "title": "Software Engineer"
                                        }
                                    }
                                ]
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Candidates"
                ],
                "summary": "Create a new candidate",
                "operationId": "1d1be0fe568ca68caf7ff18e64fd95ae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "job_id": 1,
                                "name": "John Doe",
                                "email": "john@example.com",
                                "phone": "123456789",
                                "year": 2022,
                                "skill_sets": [
                                    {
                                        "skill_id": 1
                                    },
                                    {
                                        "skill_id": 2
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Email or phone number is already associated with another candidate"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/candidates/{id}": {
            "get": {
                "tags": [
                    "Candidates"
                ],
                "summary": "Get a candidate by ID",
                "operationId": "48ec1724e465413589ad707a5b4df1cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Candidate ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "id": 1,
                                    "name": "John Doe",
                                    "email": "john@example.com",
                                    "phone": "123456789",
                                    "year": 2022,
                                    "skillSets": [
                                        {
                                            "id": 1,
                                            "skill_id": 1
                                        }
                                    ],
                                    "job": {
                                        "id": 1,
                                        "title": "Software Engineer"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Candidate not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Candidate not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Candidates"
                ],
                "summary": "Update a candidate by ID",
                "operationId": "1661cc5296116c63161ca2bf67c33a4d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Candidate ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "job_id": 1,
                                "name": "Updated John Doe",
                                "email": "updated-john@example.com",
                                "phone": "987654321",
                                "year": 2023,
                                "skill_sets": [
                                    {
                                        "skill_id": 2
                                    },
                                    {
                                        "skill_id": 3
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Candidate updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "id": 1,
                                    "name": "Updated John Doe",
                                    "email": "updated-john@example.com",
                                    "phone": "987654321",
                                    "year": 2023,
                                    "skillSets": [
                                        {
                                            "id": 2,
                                            "skill_id": 2
                                        },
                                        {
                                            "id": 3,
                                            "skill_id": 3
                                        }
                                    ],
                                    "job": {
                                        "id": 1,
                                        "title": "Software Engineer"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Candidate not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Candidate not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Candidates"
                ],
                "summary": "Delete a candidate by ID",
                "operationId": "556f8beb59b9cbd6cd49038d8d50ab19",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Candidate ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Candidate deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Candidate deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Candidate not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Candidate not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}